Refactor product scanning: remove ProductSearch component and integrate barcode scanning directly into ProductsList
